Add warning that metaplex does not exist on localhost
Heya! New metaplex user, I'm taking a guess - because it doesn't say here - that metaplex doesn't work on localhost with `solana-test-validator`. Is that correct? 

This was me after following the guide, and being used to other programs (like SPL token) that work on localhost:  https://solana.stackexchange.com/questions/1879/metaplex-create-fails-on-localhost-with-attempt-to-load-a-program-that-does-n 

So adding a note to others to make it clear they need to use another network. I hope it's useful!